Organização do site (Código)

+++ database/migrations/2019_05_27_140212_create_caracteristicas_table.php
@@ -16,7 +16,7 @@ class CreateCaracteristicasTable extends Migration
         Schema::create('caracteristicas', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('desc');
-            $table->double('preco');
+            $table->integer('preco');
             $table->timestamps();
         });
     }
